Session extended for fixed activeDuration from now
Great stuff! Thanks!

This proposal might help bring consistency on how long the session is extended from the current time.
Currently the `createdAt` is update to extend by `activeDuration`, which in-effect extends the session for a varied time from now, i.e. anywhere between `activeDuration` to `duration`. This proposal would extend the session for exactly the `activeDuration` from now.